backend/requirements.txt
fastapi
uvicorn
aiogram
aiosqlite
httpx
python-dotenv
telethon